+++ packages/flame/CHANGELOG.md
@@ -7,6 +7,7 @@
  - Add `NineTileBox.asset`
  - Generalize `paint` usage on fields
  - Create `OpacityEffect`
+ - Create `ColorEffect`
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps